+++ tasks/copy.js
@@ -2,8 +2,10 @@ const gulp = require('gulp');
 const gutil = require('gulp-util');
 const plumber = require('gulp-plumber');
 const gulpif = require('gulp-if');
-
+const copyAssets = require('npm-assets');
 const rev = require('gulp-rev');
+const glob = require('glob');
+
 const config = require('../config');
 const pathBuilder = require('../pathBuilder');
 
@@ -80,4 +82,45 @@ gulp.task('copy:img', () => {
  */
 gulp.task('copy:fonts', () => {
     copy('fonts');
+});
+
+/**
+ * `copy:assets` Task
+ * ---------------------
+ * Copy assets from from packages to the dist folder.
+ *
+ */
+gulp.task('copy:assets', callback => {
+
+    const ending = 'package.json';
+    const getPkg = filepath => {
+        const path = filepath.slice(0, -ending.length); // the parent directory that contains the package.json
+        const split = path.split('/'); // e.g. [...'@justeat', '', 'fozzie', '']
+        return {
+            path,
+            name: split[split.length - 2]
+        };
+    };
+
+    glob(config.importedAssets.importedAssetsSrcGlob, (err, files) => {
+        if (err) config.gulp.onError(err);
+
+        const process = file => new Promise((resolve, reject) => {
+            const pkg = getPkg(file);
+            gutil.log(`❯❯ Copying any assets in ${pkg.path} to ${pathBuilder.importedAssetsDistDir}/${pkg.name}`);
+            copyAssets(pkg.path, `${pathBuilder.importedAssetsDistDir}/${pkg.name}`, e => {
+                if (e) {
+                    config.gulp.onError(e);
+                    reject();
+                } else resolve();
+            });
+        });
+
+        const promises = files
+            .filter(file => file.endsWith(ending)) // Only consider folders containing a package.json
+            .map(file => process(file));
+
+        Promise.all(promises).then(callback);
+    });
+
 });