+++ tasks/css.js
@@ -31,7 +31,8 @@ const pathBuilder = require('../pathBuilder');
 gulp.task('css', callback => {
     runSequence(
         'scss:lint',
-        'clean:css',
+        ['clean:css', 'clean:assets'],
+        'copy:assets',
         'css:bundle',
         'copy:css',
         'css:lint',
@@ -116,7 +117,7 @@ gulp.task('css:bundle', () => gulp.src(`${pathBuilder.scssSrcDir}/**/*.scss`)
         postcss([
             // Converts any specified assets to data URIs
             assets({
-                loadPaths: [pathBuilder.imgSrcDir]
+                loadPaths: [pathBuilder.imgSrcDir, pathBuilder.importedAssetsDistDir]
             }),
 
             // Autoprefixes CSS properties for various browsers – browsers specified in package.json config